requirements.txt
sentence-transformers==2.4.0
langchain
flask
pypdf
transformers
torch
python-dotenv
pinecone[grpc]
langchain-pinecone
langchain_openai
langchain_community
langchain_experimental
protobuf
-e .